#21 Manage product quantity

+++ src/app/modules/shop/components/product-card/product-card.component.ts
@@ -22,13 +22,14 @@ export class ProductCardComponent {
   }
 
   addProductToCart() {
-    if (this.product.quantity && this.product.quantity > 0) return
+    if ((this.product.quantity && this.product.quantity > 0) || this.product.quantity === this.product.AvailablePieces) return
     this.product.quantity = this.quantity = 1
     this.cartService.addProduct(this.product)
   }
 
   editProductQuantityInCart(operation: 'add' | 'subtract' | 'change') {
     if (operation === 'add') {
+      if (this.product.quantity === this.product.AvailablePieces) return
       this.quantity += 1
       this.product.quantity = this.product.quantity ? this.product.quantity + 1 : 1
       this.cartService.editProductQuantityInCart(this.product)
@@ -40,6 +41,11 @@ export class ProductCardComponent {
       else this.cartService.removeProductFromCart(this.product)
     } 
     else if (operation === 'change') {
+      if (this.product.quantity === this.product.AvailablePieces) {
+        this.quantity = this.product.quantity
+        return
+      }
+
       this.product.quantity = this.quantity
       if (this.quantity > 0) this.cartService.editProductQuantityInCart(this.product)
       else this.cartService.removeProductFromCart(this.product)
